+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - **Textarea**, **Textfield**: dispatching native `input` and `change` events.
 - **Radio**, **Checkbox**, **SingleSelect**, **MultiSelect**: dispatching native `change` event
 - **Collapsible**: Displaying the main content using the default slot instead of the named `body` slot.
-- **Split Layout**: Renamed `initial-pos` to `initial-position`.
+- **Split Layout**: Renamed `initial-pos` to `initial-handle-position`.
 
 ### Removed